Sensors to pd
Created 2 abstractions to get proxi and adc datas and connect it to the synth.

Added a bit of arduino code to get the values of the sensors (missing teensy's adc for the moment).

N.B.: The adc i get for the moment is the one from the computer, not the teensy.